refactor: remove legacy settings dialog code
