--- app/src/main/feature/settings/drivers/DXVKConfigDialog.java
@@ -1,312 +1,0 @@
-package com.winlator.cmod.feature.settings;
-
-import android.content.Context;
-import android.util.Log;
-import android.view.View;
-import android.widget.AdapterView;
-import android.widget.CompoundButton;
-import android.widget.Spinner;
-
-import com.winlator.cmod.R;
-import com.winlator.cmod.runtime.container.Container;
-import com.winlator.cmod.runtime.content.ContentProfile;
-import com.winlator.cmod.runtime.content.ContentsManager;
-import com.winlator.cmod.shared.android.AppUtils;
-import com.winlator.cmod.shared.ui.dialog.ContentDialog;
-import com.winlator.cmod.runtime.wine.DefaultVersion;
-import com.winlator.cmod.runtime.wine.EnvVars;
-import com.winlator.cmod.shared.util.KeyValueSet;
-import com.winlator.cmod.shared.util.StringUtils;
-import com.winlator.cmod.runtime.wine.VKD3DVersionItem;
-import com.winlator.cmod.runtime.display.environment.ImageFs;
-
-import java.io.File;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
-
-public class DXVKConfigDialog extends ContentDialog {
-    public static final String DEFAULT_CONFIG = Container.DEFAULT_DXWRAPPERCONFIG;
-    public static final int DXVK_TYPE_NONE = 0;
-    public static final int DXVK_TYPE_ASYNC = 1;
-    public static final int DXVK_TYPE_GPLASYNC = 2;
-    private final CompoundButton swAsync;
-    private boolean isARM64EC = false;
-    private final CompoundButton swAsyncCache;
-    private final View llAsync;
-    private final View llAsyncCache;
-    private final Context context;
-    private static List<String> dxvkVersions;
-    private static final Pattern SEMVER = Pattern.compile("(\\d+)\\.(\\d+)(?:\\.(\\d+))?");
-
-    private static Integer tryGetMajor(String s) {
-        if (s == null) return null;
-        Matcher m = SEMVER.matcher(s);
-        if (!m.find()) return null;
-        try {
-            return Integer.parseInt(m.group(1));
-        } catch (NumberFormatException e) {
-            return null;
-        }
-    }
-    
-    public static final String[] VKD3D_FEATURE_LEVEL = {"12_0", "12_1", "12_2", "11_1", "11_0", "10_1", "10_0", "9_3", "9_2", "9_1"};
-
-    private static int compareVersion(String varA, String varB) {
-        final String[] levelsA = varA.split("\\.");
-        final String[] levelsB = varB.split("\\.");
-        int minLen = Math.min(levelsA.length, levelsB.length);
-        int numA, numB;
-
-        for (int i = 0; i < minLen; i++) {
-            numA = Integer.parseInt(levelsA[i]);
-            numB = Integer.parseInt(levelsB[i]);
-            if (numA != numB)
-                return numA - numB;
-        }
-
-        if (levelsA.length != levelsB.length)
-            return levelsA.length - levelsB.length;
-
-        return 0;
-    }
-
-    public DXVKConfigDialog(View anchor, boolean isARM64EC) {
-        super(anchor.getContext(), R.layout.dxvk_config_dialog);
-        context = anchor.getContext();
-        setIcon(R.drawable.icon_settings);
-        setTitle("DXVK "+context.getString(R.string.container_config_title));
-
-        final Spinner sDXVKVersion = findViewById(R.id.SDXVKVersion);
-        final Spinner sVKD3DVersion = findViewById(R.id.SVKD3DVersion);
-        final Spinner sFramerate = findViewById(R.id.SFramerate);
-        final Spinner sVKD3DFeatureLevel = findViewById(R.id.SVKD3DFeatureLevel);
-        final Spinner sDDRAWrapper = findViewById(R.id.SDDRAWrapper);
-        swAsync = (CompoundButton) findViewById(R.id.SWAsync);
-        swAsyncCache = (CompoundButton) findViewById(R.id.SWAsyncCache);
-        llAsync = findViewById(R.id.LLAsync);
-        llAsyncCache = findViewById(R.id.LLAsyncCache);
-
-        ContentsManager contentsManager = new ContentsManager(context);
-        contentsManager.syncContents();
-
-        KeyValueSet config = parseConfig(anchor.getTag());
-        loadDxvkVersionSpinner(contentsManager, sDXVKVersion, isARM64EC);
-        loadVkd3dVersionSpinner(contentsManager, sVKD3DVersion);
-
-        AppUtils.setupThemedSpinner(sVKD3DFeatureLevel, context, Arrays.asList(VKD3D_FEATURE_LEVEL));
-        AppUtils.setupThemedSpinner(sDDRAWrapper, context, Arrays.asList(context.getResources().getStringArray(R.array.ddrawrapper_entries)));
-        AppUtils.setupThemedSpinner(sFramerate, context, Arrays.asList(context.getResources().getStringArray(R.array.dxvk_framerate_entries)));
-
-        setDXVKSpinner(sDXVKVersion, config, contentsManager, isARM64EC);
-        AppUtils.setSpinnerSelectionFromIdentifier(sFramerate, config.get("framerate"));
-        AppUtils.setSpinnerSelectionFromIdentifier(sVKD3DVersion, config.get("vkd3dVersion"));
-        AppUtils.setSpinnerSelectionFromIdentifier(sVKD3DFeatureLevel, config.get("vkd3dLevel"));
-        AppUtils.setSpinnerSelectionFromIdentifier(sDDRAWrapper, config.get("ddrawrapper"));
-
-        swAsync.setChecked(config.get("async").equals("1"));
-        swAsyncCache.setChecked(config.get("asyncCache").equals("1"));
-
-        updateConfigVisibility(getDXVKType(sDXVKVersion.getSelectedItemPosition()));
-
-        sDXVKVersion.setOnItemSelectedListener(new AdapterView.OnItemSelectedListener() {
-            @Override
-            public void onItemSelected(AdapterView<?> parent, View view, int position, long id) {
-                updateConfigVisibility(getDXVKType(position));
-            }
-
-            @Override
-            public void onNothingSelected(AdapterView<?> parent) {
-
-            }
-        });
-
-        sVKD3DVersion.setOnItemSelectedListener(new AdapterView.OnItemSelectedListener() {
-            @Override
-            public void onItemSelected(AdapterView<?> parent, View view, int position, long id) {
-                Object selectedItem = sVKD3DVersion.getSelectedItem();
-                String selectedVersion = selectedItem != null ? selectedItem.toString() : "None";
-                String currentDXVKVersion = config.get("version");
-
-                if (!selectedVersion.equals("None")) {
-                    ArrayList<String> filteredVersions = new ArrayList<>();
-
-                    for (int i = 0; i < dxvkVersions.size(); i++) {
-                        Integer major = tryGetMajor(dxvkVersions.get(i));
-                        if (major == null || major >= 2) {
-                            filteredVersions.add(dxvkVersions.get(i));
-                        }
-                    }
-
-                    AppUtils.setupThemedSpinner(sDXVKVersion, context, filteredVersions);
-
-                    Integer curMajor = tryGetMajor(currentDXVKVersion);
-                    AppUtils.setSpinnerSelectionFromIdentifier(
-                            sDXVKVersion,
-                            (curMajor != null && curMajor >= 2) ? currentDXVKVersion : DefaultVersion.DXVK
-                    );
-                    updateConfigVisibility(getDXVKType(sDXVKVersion.getSelectedItemPosition()));
-                }
-                else {
-                    loadDxvkVersionSpinner(contentsManager, sDXVKVersion, isARM64EC);
-                    AppUtils.setSpinnerSelectionFromIdentifier(sDXVKVersion, config.get("version"));
-                }
-            }
-
-            @Override
-            public void onNothingSelected(AdapterView<?> parent) {
-            }
-        });
-
-        setOnConfirmCallback(() -> {
-            config.put("version", sDXVKVersion.getSelectedItem() != null ? sDXVKVersion.getSelectedItem().toString() : DefaultVersion.DXVK);
-            config.put("framerate", sFramerate.getSelectedItem() != null ? StringUtils.parseNumber(sFramerate.getSelectedItem()) : "0");
-            config.put("async", ((swAsync.isChecked())&&(llAsync.getVisibility()==View.VISIBLE))?"1":"0");
-            config.put("asyncCache", ((swAsyncCache.isChecked())&&(llAsyncCache.getVisibility()==View.VISIBLE))?"1":"0");
-            Object vkd3dItem = sVKD3DVersion.getSelectedItem();
-            if (vkd3dItem instanceof VKD3DVersionItem) {
-                config.put("vkd3dVersion", ((VKD3DVersionItem) vkd3dItem).getIdentifier());
-            } else {
-                config.put("vkd3dVersion", vkd3dItem != null ? vkd3dItem.toString() : "None");
-            }
-            config.put("vkd3dLevel", sVKD3DFeatureLevel.getSelectedItem() != null ? sVKD3DFeatureLevel.getSelectedItem().toString() : "12_0");
-            config.put("ddrawrapper", sDDRAWrapper.getSelectedItem() != null ? StringUtils.parseIdentifier(sDDRAWrapper.getSelectedItem().toString()) : Container.DEFAULT_DDRAWRAPPER);
-            anchor.setTag(config.toString());
-        });
-    }
-
-    private void updateConfigVisibility(int dxvkType) {
-        if (dxvkType == DXVK_TYPE_ASYNC) {
-            llAsync.setVisibility(View.VISIBLE);
-            llAsyncCache.setVisibility(View.GONE);
-        } else if (dxvkType == DXVK_TYPE_GPLASYNC) {
-            llAsync.setVisibility(View.VISIBLE);
-            llAsyncCache.setVisibility(View.VISIBLE);
-        } else {
-            llAsync.setVisibility(View.GONE);
-            llAsyncCache.setVisibility(View.GONE);
-        }
-    }
-
-    private int getDXVKType(int pos) {
-        if (dxvkVersions == null || pos < 0 || pos >= dxvkVersions.size()) return DXVK_TYPE_NONE;
-        final String v = dxvkVersions.get(pos);
-        int dxvkType = DXVK_TYPE_NONE;
-        if (v.contains("gplasync"))
-            dxvkType = DXVK_TYPE_GPLASYNC;
-        else if (v.contains("async"))
-            dxvkType = DXVK_TYPE_ASYNC;
-        return dxvkType;
-    }
-
-    private void setDXVKSpinner(Spinner sDXVKVersion, KeyValueSet config, ContentsManager contentsManager, boolean isARM64EC) {
-        String selectedVersion = config.get("vkd3dVersion");
-        String currentDXVKVersion = config.get("version");
-        if (!selectedVersion.equals("None")) {
-            ArrayList<String> filteredVersions = new ArrayList<>();
-
-            for (int i = 0; i < dxvkVersions.size(); i++) {
-                Integer major = tryGetMajor(dxvkVersions.get(i));
-                if (major == null || major >= 2) {
-                    filteredVersions.add(dxvkVersions.get(i));
-                }
-            }
-
-            AppUtils.setupThemedSpinner(sDXVKVersion, context, filteredVersions);
-
-            Integer curMajor = tryGetMajor(currentDXVKVersion);
-            AppUtils.setSpinnerSelectionFromIdentifier(
-                    sDXVKVersion,
-                    (curMajor != null && curMajor >= 2) ? currentDXVKVersion : DefaultVersion.DXVK
-            );
-        }
-        else
-            AppUtils.setSpinnerSelectionFromIdentifier(sDXVKVersion, currentDXVKVersion);
-    }
-
-    public static KeyValueSet parseConfig(Object config) {
-        String data = config != null && !config.toString().isEmpty() ? config.toString() :  DEFAULT_CONFIG;
-        return new KeyValueSet(data);
-    }
-
-    public static void setEnvVars(Context context, KeyValueSet config, EnvVars envVars) {
-        setEnvVars(context, config, envVars, 0);
-    }
-
-    public static void setEnvVars(Context context, KeyValueSet config, EnvVars envVars, int refreshRateOverride) {
-        String content = "";
-
-        // Refresh rate override takes precedence over per-container DXVK framerate
-        if (refreshRateOverride > 0) {
-            String rateStr = String.valueOf(refreshRateOverride);
-            content += "dxgi.syncInterval = 0; ";
-            content += "dxgi.maxFrameRate = " + rateStr + "; ";
-            content += "d3d9.maxFrameRate = " + rateStr;
-            envVars.put("DXVK_FRAME_RATE", rateStr);
-        } else {
-            String framerate = config.get("framerate");
-
-            if (!framerate.isEmpty() && !framerate.equals("0")) {
-                content += "dxgi.maxFrameRate = " + framerate + "; ";
-                content += "d3d9.maxFrameRate = " + framerate;
-                envVars.put("DXVK_FRAME_RATE", framerate);
-            }
-        }
-
-        String async = config.get("async");
-        if (!async.isEmpty() && !async.equals("0"))
-            envVars.put("DXVK_ASYNC", "1");
-
-        String asyncCache = config.get("asyncCache");
-        if (!asyncCache.isEmpty() && !asyncCache.equals("0"))
-            envVars.put("DXVK_GPLASYNCCACHE", "1");
-
-        if (!content.isEmpty())
-            envVars.put("DXVK_CONFIG", content);
-
-        envVars.put("VKD3D_FEATURE_LEVEL", config.get("vkd3dLevel"));
-        envVars.put("DXVK_STATE_CACHE_PATH", context.getFilesDir() + "/imagefs/" + ImageFs.CACHE_PATH);
-    }
-
-    private void loadDxvkVersionSpinner(ContentsManager manager, Spinner spinner, boolean isARM64EC) {
-        this.isARM64EC = isARM64EC;
-        String[] originalItems = context.getResources().getStringArray(R.array.dxvk_version_entries);
-        List<String> itemList = new ArrayList<>(Arrays.asList(originalItems));
-
-        for (ContentProfile profile : manager.getProfiles(ContentProfile.ContentType.CONTENT_TYPE_DXVK)) {
-            String entryName = ContentsManager.getEntryName(profile);
-            int firstDashIndex = entryName.indexOf('-');
-            itemList.add(entryName.substring(firstDashIndex + 1));
-        }
-
-        // Remove arm64ec items using reverse iteration to avoid skipping elements
-        for (int i = itemList.size() - 1; i >= 0; i--) {
-            if (itemList.get(i).contains("arm64ec") && !isARM64EC)
-                itemList.remove(i);
-        }
-
-        AppUtils.setupThemedSpinner(spinner, context, itemList);
-        dxvkVersions = new ArrayList<>(itemList);
-    }
-
-    private void loadVkd3dVersionSpinner(ContentsManager manager, Spinner spinner) {
-        List<VKD3DVersionItem> itemList = new ArrayList<>();
-
-        // Add predefined versions
-        String[] originalItems = context.getResources().getStringArray(R.array.vkd3d_version_entries);
-        for (String version : originalItems) {
-            itemList.add(new VKD3DVersionItem(version)); // For predefined versions, use 0 as verCode
-        }
-
-        // Add installed content profiles
-        for (ContentProfile profile : manager.getProfiles(ContentProfile.ContentType.CONTENT_TYPE_VKD3D)) {
-            String displayName = profile.verName;  // Display name for the spinner
-            int versionCode = profile.verCode;     // Unique version code if available
-            itemList.add(new VKD3DVersionItem(displayName, versionCode));
-        }
-
-        AppUtils.setupThemedSpinner(spinner, context, itemList);
-    }
-}
