--- app/src/main/res/layout/add_env_var_dialog.xml
@@ -1,41 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="300dp"
-    android:layout_height="wrap_content"
-    android:orientation="vertical"
-    android:paddingTop="2dp"
-    android:paddingBottom="4dp">
-
-    <TextView
-        style="@style/OtherSettingsFieldLabel"
-        android:text="@string/common_ui_name" />
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginBottom="10dp"
-        android:gravity="center_vertical"
-        android:orientation="horizontal">
-
-        <EditText
-            style="@style/OtherSettingsEditText"
-            android:id="@+id/ETName"
-            android:layout_width="0dp"
-            android:layout_weight="1" />
-
-        <ImageButton
-            style="@style/OtherSettingsIconButton"
-            android:id="@+id/BTMenu"
-            android:layout_marginStart="6dp"
-            android:src="@drawable/icon_expand_more" />
-    </LinearLayout>
-
-    <TextView
-        style="@style/OtherSettingsFieldLabel"
-        android:text="@string/common_ui_value" />
-
-    <EditText
-        style="@style/OtherSettingsEditText"
-        android:id="@+id/ETValue"
-        android:layout_width="match_parent" />
-</LinearLayout>
