--- app/src/main/res/layout/container_storage_info_dialog.xml
@@ -1,128 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<ScrollView xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="172dp"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
-
-    <LinearLayout
-        android:layout_width="330dp"
-        android:layout_height="wrap_content"
-        android:orientation="horizontal"
-        android:padding="8dp"
-        android:gravity="center_vertical">
-
-        <LinearLayout
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:layout_weight="1"
-            android:orientation="vertical">
-
-            <TextView
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:fontFamily="@font/inter"
-                android:textColor="@color/settings_text_secondary"
-                android:textSize="11sp"
-                android:letterSpacing="0.06"
-                android:textAllCaps="true"
-                android:textStyle="bold"
-                android:text="@string/container_config_drive_c" />
-
-            <TextView
-                android:id="@+id/TVDriveCSize"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:fontFamily="@font/inter"
-                android:textColor="@color/settings_accent"
-                android:textSize="15sp"
-                android:textStyle="bold"
-                android:layout_marginTop="2dp" />
-
-            <TextView
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:fontFamily="@font/inter"
-                android:textColor="@color/settings_text_secondary"
-                android:textSize="11sp"
-                android:letterSpacing="0.06"
-                android:textAllCaps="true"
-                android:textStyle="bold"
-                android:text="@string/container_config_cache"
-                android:layout_marginTop="10dp" />
-
-            <TextView
-                android:id="@+id/TVCacheSize"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:fontFamily="@font/inter"
-                android:textColor="@color/settings_accent"
-                android:textSize="15sp"
-                android:textStyle="bold"
-                android:layout_marginTop="2dp" />
-
-            <TextView
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:fontFamily="@font/inter"
-                android:textColor="@color/settings_text_secondary"
-                android:textSize="11sp"
-                android:letterSpacing="0.06"
-                android:textAllCaps="true"
-                android:textStyle="bold"
-                android:text="@string/container_config_total"
-                android:layout_marginTop="10dp" />
-
-            <TextView
-                android:id="@+id/TVTotalSize"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:fontFamily="@font/inter"
-                android:textColor="@color/settings_accent"
-                android:textSize="15sp"
-                android:textStyle="bold"
-                android:layout_marginTop="2dp" />
-        </LinearLayout>
-
-        <LinearLayout
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:layout_weight="1"
-            android:gravity="center_horizontal"
-            android:orientation="vertical">
-
-            <FrameLayout
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content">
-
-                <com.google.android.material.progressindicator.CircularProgressIndicator
-                    android:id="@+id/CircularProgressIndicator"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    app:trackThickness="12dp"
-                    app:trackColor="@color/settings_outline"
-                    app:indicatorColor="@color/settings_accent"
-                    app:indicatorSize="100dp"
-                    android:layout_gravity="center" />
-
-                <TextView
-                    android:id="@+id/TVUsedSpace"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:layout_gravity="center"
-                    android:fontFamily="@font/inter"
-                    android:textColor="@color/settings_text_primary"
-                    android:textSize="15sp"
-                    android:textStyle="bold" />
-            </FrameLayout>
-
-            <TextView
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:fontFamily="@font/inter"
-                android:textSize="11sp"
-                android:layout_marginTop="6dp"
-                android:textColor="@color/settings_text_secondary"
-                android:text="@string/container_config_estimated_used_space" />
-        </LinearLayout>
-    </LinearLayout>
-</ScrollView>
