--- app/src/main/res/layout/dxvk_config_dialog.xml
@@ -1,141 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:layout_width="288dp"
-    android:layout_height="wrap_content"
-    android:orientation="vertical"
-    android:paddingTop="2dp"
-    android:paddingBottom="4dp">
-
-    <androidx.core.widget.NestedScrollView
-        android:layout_width="match_parent"
-        android:layout_height="320dp"
-        android:clipToPadding="false">
-
-        <LinearLayout
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:orientation="vertical"
-            android:paddingEnd="8dp"
-            android:paddingBottom="4dp">
-
-            <TextView
-                style="@style/OtherSettingsFieldLabel"
-                android:text="@string/container_wine_vkd3d_version" />
-
-            <androidx.appcompat.widget.AppCompatSpinner
-                android:id="@+id/SVKD3DVersion"
-                style="@style/OtherSettingsSpinner"
-                android:layout_width="match_parent"
-                android:layout_marginTop="4dp"
-                android:spinnerMode="dropdown" />
-
-            <TextView
-                style="@style/OtherSettingsFieldLabel"
-                android:layout_marginTop="10dp"
-                android:text="@string/container_wine_vkd3d_feature_level" />
-
-            <androidx.appcompat.widget.AppCompatSpinner
-                android:id="@+id/SVKD3DFeatureLevel"
-                style="@style/OtherSettingsSpinner"
-                android:layout_width="match_parent"
-                android:layout_marginTop="4dp"
-                android:spinnerMode="dropdown" />
-
-            <TextView
-                style="@style/OtherSettingsFieldLabel"
-                android:layout_marginTop="10dp"
-                android:text="@string/container_wine_dxvk_version" />
-
-            <androidx.appcompat.widget.AppCompatSpinner
-                android:id="@+id/SDXVKVersion"
-                style="@style/OtherSettingsSpinner"
-                android:layout_width="match_parent"
-                android:layout_marginTop="4dp"
-                android:spinnerMode="dropdown" />
-
-            <LinearLayout
-                android:id="@+id/LLAsync"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:layout_marginTop="10dp"
-                android:background="@drawable/env_vars_row_background"
-                android:gravity="center_vertical"
-                android:orientation="horizontal"
-                android:paddingHorizontal="10dp"
-                android:paddingVertical="8dp">
-
-                <TextView
-                    android:layout_width="0dp"
-                    android:layout_height="wrap_content"
-                    android:layout_weight="1"
-                    android:fontFamily="@font/inter"
-                    android:text="@string/container_wine_enabled_async"
-                    android:textColor="@color/settings_text_primary"
-                    android:textSize="13sp" />
-
-                <com.google.android.material.switchmaterial.SwitchMaterial
-                    android:id="@+id/SWAsync"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:background="@android:color/transparent"
-                    app:thumbTint="@color/settings_accent"
-                    app:trackTintMode="src_in" />
-            </LinearLayout>
-
-            <LinearLayout
-                android:id="@+id/LLAsyncCache"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:layout_marginTop="6dp"
-                android:background="@drawable/env_vars_row_background"
-                android:gravity="center_vertical"
-                android:orientation="horizontal"
-                android:paddingHorizontal="10dp"
-                android:paddingVertical="8dp">
-
-                <TextView
-                    android:layout_width="0dp"
-                    android:layout_height="wrap_content"
-                    android:layout_weight="1"
-                    android:fontFamily="@font/inter"
-                    android:text="@string/container_wine_enabled_async_cache"
-                    android:textColor="@color/settings_text_primary"
-                    android:textSize="13sp" />
-
-                <com.google.android.material.switchmaterial.SwitchMaterial
-                    android:id="@+id/SWAsyncCache"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:background="@android:color/transparent"
-                    app:thumbTint="@color/settings_accent"
-                    app:trackTintMode="src_in" />
-            </LinearLayout>
-
-            <TextView
-                style="@style/OtherSettingsFieldLabel"
-                android:layout_marginTop="10dp"
-                android:text="@string/container_wine_ddraw_wrapper" />
-
-            <androidx.appcompat.widget.AppCompatSpinner
-                android:id="@+id/SDDRAWrapper"
-                style="@style/OtherSettingsSpinner"
-                android:layout_width="match_parent"
-                android:layout_marginTop="4dp"
-                android:spinnerMode="dropdown" />
-
-            <TextView
-                style="@style/OtherSettingsFieldLabel"
-                android:layout_marginTop="10dp"
-                android:text="@string/session_display_frame_rate" />
-
-            <androidx.appcompat.widget.AppCompatSpinner
-                android:id="@+id/SFramerate"
-                style="@style/OtherSettingsSpinner"
-                android:layout_width="match_parent"
-                android:layout_marginTop="4dp"
-                android:spinnerMode="dropdown" />
-
-        </LinearLayout>
-    </androidx.core.widget.NestedScrollView>
-</LinearLayout>
